# server/swagger.yaml
openapi: 3.0.0
info:
  title: Motio API
  version: 1.0.0
  description: API for Motio Car Dealership Logistics
servers:
  - url: /api
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
...
security:
  - cookieAuth: []
paths:
  /auth/login:
    post:
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: OK
...